+++ anvay/anvay/src/main/java/com/cts/mfrp/anvay/service/ApplicationService.java
@@ -1,13 +1,12 @@
 package com.cts.mfrp.anvay.service;
 
-import java.util.List;
-
 import com.cts.mfrp.anvay.entity.LeadershipApplication;
+import java.util.List;
 
 public interface ApplicationService {
     LeadershipApplication createApplication(LeadershipApplication application);
-    LeadershipApplication getApplicationById(Integer applicationId);
-    List<LeadershipApplication> getApplicationsByClubId(Integer clubId);
-    LeadershipApplication updateApplication(Integer applicationId, LeadershipApplication application);
-    void deleteApplication(Integer applicationId);
+    LeadershipApplication getApplicationById(Long applicationId);
+    List<LeadershipApplication> getApplicationsByClubId(Long clubId);
+    LeadershipApplication updateApplication(Long applicationId, LeadershipApplication application);
+    void deleteApplication(Long applicationId);
 }
